Trigger BufReadPost at completion of search.
Set the QuickFix/Location list at the completion of the search.  This
triggers the BufReadPost autocommand that some plugins rely on to
perform post-processing of the search results.  Appending to a
QuickFix/Location list does not send the BufReadPost, but setting the
list does (even if the new value is the same as the current value).